move error classes to a seperate file
